More impala-sqlalchemy separation code refactoring / cleaning
