+++ ibis/impala/ddl.py
@@ -15,8 +15,8 @@
 from ibis.compat import StringIO
 import re
 
-from ibis.sql.ddl import DDL, Select
-from .exprs import quote_identifier, _type_to_sql_string
+from ibis.sql.compiler import DDL
+from .compiler import quote_identifier, _type_to_sql_string
 
 from ibis.expr.datatypes import validate_type
 from ibis.compat import py_string
@@ -51,10 +51,6 @@ class ImpalaDDL(DDL):
             else:
                 return obj_name
         return scoped_name
-
-
-class ImpalaSelect(Select):
-    pass
 
 
 class CreateDDL(ImpalaDDL):
