+++ ibis/impala/tests/common.py
@@ -143,7 +143,7 @@ class ImpalaE2E(object):
 
 
 def format_schema(expr):
-    from ibis.impala.exprs import _type_to_sql_string
+    from ibis.impala.compiler import _type_to_sql_string
     from pprint import pprint
     schema = expr.schema()
