--- ibis/sql/ddl.py
@@ -1,536 +1,0 @@
-# Copyright 2014 Cloudera Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from ibis.compat import StringIO
-
-import ibis.expr.types as ir
-import ibis.expr.operations as ops
-import ibis.common as com
-import ibis.util as util
-
-
-class DDL(object):
-    pass
-
-
-class Select(DDL):
-
-    """
-    A SELECT statement which, after execution, might yield back to the user a
-    table, array/list, or scalar value, depending on the expression that
-    generated it
-    """
-
-    def __init__(self, table_set, select_set,
-                 subqueries=None, where=None, group_by=None, having=None,
-                 order_by=None, limit=None,
-                 distinct=False, indent=2,
-                 result_handler=None, parent_expr=None,
-                 context=None):
-        self.context = context
-
-        self.select_set = select_set
-        self.table_set = table_set
-        self.distinct = distinct
-
-        self.parent_expr = parent_expr
-
-        self.where = where or []
-
-        # Group keys and post-predicates for aggregations
-        self.group_by = group_by or []
-        self.having = having or []
-        self.order_by = order_by or []
-
-        self.limit = limit
-        self.subqueries = subqueries or []
-
-        self.indent = indent
-
-        self.result_handler = result_handler
-
-    def equals(self, other):
-        if not isinstance(other, Select):
-            return False
-
-        this_exprs = self._all_exprs()
-        other_exprs = other._all_exprs()
-
-        if self.limit != other.limit:
-            return False
-
-        for x, y in zip(this_exprs, other_exprs):
-            if not x.equals(y):
-                return False
-
-        return True
-
-    def _all_exprs(self):
-        # Gnarly, maybe we can improve this somehow
-        expr_attrs = ['select_set', 'table_set', 'where', 'group_by', 'having',
-                      'order_by', 'subqueries']
-        exprs = []
-        for attr in expr_attrs:
-            val = getattr(self, attr)
-            if isinstance(val, list):
-                exprs.extend(val)
-            else:
-                exprs.append(val)
-
-        return exprs
-
-    def compile(self):
-        """
-        This method isn't yet idempotent; calling multiple times may yield
-        unexpected results
-        """
-        # Can't tell if this is a hack or not. Revisit later
-        self.context.set_query(self)
-
-        # If any subqueries, translate them and add to beginning of query as
-        # part of the WITH section
-        with_frag = self.format_subqueries()
-
-        # SELECT
-        select_frag = self.format_select_set()
-
-        # FROM, JOIN, UNION
-        from_frag = self.format_table_set()
-
-        # WHERE
-        where_frag = self.format_where()
-
-        # GROUP BY and HAVING
-        groupby_frag = self.format_group_by()
-
-        # ORDER BY and LIMIT
-        order_frag = self.format_postamble()
-
-        # Glue together the query fragments and return
-        query = _join_not_none('\n', [with_frag, select_frag, from_frag,
-                                      where_frag, groupby_frag, order_frag])
-
-        return query
-
-    def format_subqueries(self):
-        if len(self.subqueries) == 0:
-            return
-
-        context = self.context
-
-        buf = StringIO()
-        buf.write('WITH ')
-
-        for i, expr in enumerate(self.subqueries):
-            if i > 0:
-                buf.write(',\n')
-            formatted = util.indent(context.get_compiled_expr(expr), 2)
-            alias = context.get_ref(expr)
-            buf.write('{0} AS (\n{1}\n)'.format(alias, formatted))
-
-        return buf.getvalue()
-
-    def format_select_set(self):
-        # TODO:
-        context = self.context
-        formatted = []
-        for expr in self.select_set:
-            if isinstance(expr, ir.ValueExpr):
-                expr_str = self._translate(expr, named=True)
-            elif isinstance(expr, ir.TableExpr):
-                # A * selection, possibly prefixed
-                if context.need_aliases():
-                    alias = context.get_ref(expr)
-
-                    # materialized join will not have an alias. see #491
-                    expr_str = '{0}.*'.format(alias) if alias else '*'
-                else:
-                    expr_str = '*'
-            formatted.append(expr_str)
-
-        buf = StringIO()
-        line_length = 0
-        max_length = 70
-        tokens = 0
-        for i, val in enumerate(formatted):
-            # always line-break for multi-line expressions
-            if val.count('\n'):
-                if i:
-                    buf.write(',')
-                buf.write('\n')
-                indented = util.indent(val, self.indent)
-                buf.write(indented)
-
-                # set length of last line
-                line_length = len(indented.split('\n')[-1])
-                tokens = 1
-            elif (tokens > 0 and line_length and
-                  len(val) + line_length > max_length):
-                # There is an expr, and adding this new one will make the line
-                # too long
-                buf.write(',\n       ') if i else buf.write('\n')
-                buf.write(val)
-                line_length = len(val) + 7
-                tokens = 1
-            else:
-                if i:
-                    buf.write(',')
-                buf.write(' ')
-                buf.write(val)
-                tokens += 1
-                line_length += len(val) + 2
-
-        if self.distinct:
-            select_key = 'SELECT DISTINCT'
-        else:
-            select_key = 'SELECT'
-
-        return '{0}{1}'.format(select_key, buf.getvalue())
-
-    def format_table_set(self):
-        if self.table_set is None:
-            return None
-
-        fragment = 'FROM '
-
-        helper = _TableSetFormatter(self, self.table_set)
-        fragment += helper.get_result()
-
-        return fragment
-
-    def format_group_by(self):
-        if not len(self.group_by):
-            # There is no aggregation, nothing to see here
-            return None
-
-        lines = []
-        if len(self.group_by) > 0:
-            clause = 'GROUP BY {0}'.format(', '.join([
-                str(x + 1) for x in self.group_by]))
-            lines.append(clause)
-
-        if len(self.having) > 0:
-            trans_exprs = []
-            for expr in self.having:
-                translated = self._translate(expr)
-                trans_exprs.append(translated)
-            lines.append('HAVING {0}'.format(' AND '.join(trans_exprs)))
-
-        return '\n'.join(lines)
-
-    def format_where(self):
-        if len(self.where) == 0:
-            return None
-
-        buf = StringIO()
-        buf.write('WHERE ')
-        fmt_preds = [self._translate(pred, permit_subquery=True)
-                     for pred in self.where]
-        conj = ' AND\n{0}'.format(' ' * 6)
-        buf.write(conj.join(fmt_preds))
-        return buf.getvalue()
-
-    def format_postamble(self):
-        buf = StringIO()
-        lines = 0
-
-        if len(self.order_by) > 0:
-            buf.write('ORDER BY ')
-            formatted = []
-            for expr in self.order_by:
-                key = expr.op()
-                translated = self._translate(key.expr)
-                if not key.ascending:
-                    translated += ' DESC'
-                formatted.append(translated)
-            buf.write(', '.join(formatted))
-            lines += 1
-
-        if self.limit is not None:
-            if lines:
-                buf.write('\n')
-            n, offset = self.limit['n'], self.limit['offset']
-            buf.write('LIMIT {0}'.format(n))
-            if offset is not None and offset != 0:
-                buf.write(' OFFSET {0}'.format(offset))
-            lines += 1
-
-        if not lines:
-            return None
-
-        return buf.getvalue()
-
-    def _translate(self, expr, context=None, named=False,
-                   permit_subquery=False):
-
-        if context is None:
-            context = self.context
-
-        translator = self.translator(expr, context=context,
-                                     named=named,
-                                     permit_subquery=permit_subquery)
-        return translator.get_result()
-
-    @property
-    def translator(self):
-        from ibis.impala.exprs import ImpalaExprTranslator
-        return ImpalaExprTranslator
-
-
-class _TableSetFormatter(object):
-    _join_names = {
-        ops.InnerJoin: 'INNER JOIN',
-        ops.LeftJoin: 'LEFT OUTER JOIN',
-        ops.RightJoin: 'RIGHT OUTER JOIN',
-        ops.OuterJoin: 'FULL OUTER JOIN',
-        ops.LeftAntiJoin: 'LEFT ANTI JOIN',
-        ops.LeftSemiJoin: 'LEFT SEMI JOIN',
-        ops.CrossJoin: 'CROSS JOIN'
-    }
-
-    def __init__(self, parent, expr, indent=2):
-        self.parent = parent
-        self.context = parent.context
-        self.expr = expr
-        self.indent = indent
-
-        self.join_tables = []
-        self.join_types = []
-        self.join_predicates = []
-
-    def get_result(self):
-        # Got to unravel the join stack; the nesting order could be
-        # arbitrary, so we do a depth first search and push the join tokens
-        # and predicates onto a flat list, then format them
-        op = self.expr.op()
-
-        if isinstance(op, ops.Join):
-            self._walk_join_tree(op)
-        else:
-            self.join_tables.append(self._format_table(self.expr))
-
-        # TODO: Now actually format the things
-        buf = StringIO()
-        buf.write(self.join_tables[0])
-        for jtype, table, preds in zip(self.join_types, self.join_tables[1:],
-                                       self.join_predicates):
-            buf.write('\n')
-            buf.write(util.indent('{0} {1}'.format(jtype, table), self.indent))
-
-            if len(preds):
-                buf.write('\n')
-                fmt_preds = [self._translate(pred) for pred in preds]
-                conj = ' AND\n{0}'.format(' ' * 3)
-                fmt_preds = util.indent('ON ' + conj.join(fmt_preds),
-                                        self.indent * 2)
-                buf.write(fmt_preds)
-
-        return buf.getvalue()
-
-    def _translate(self, expr):
-        return self.parent._translate(expr, context=self.context)
-
-    def _walk_join_tree(self, op):
-        left = op.left.op()
-        right = op.right.op()
-
-        if util.all_of([left, right], ops.Join):
-            raise NotImplementedError('Do not support joins between '
-                                      'joins yet')
-
-        self._validate_join_predicates(op.predicates)
-
-        jname = self._get_join_type(op)
-
-        # Read off tables and join predicates left-to-right in
-        # depth-first order
-        if isinstance(left, ops.Join):
-            self._walk_join_tree(left)
-            self.join_tables.append(self._format_table(op.right))
-            self.join_types.append(jname)
-            self.join_predicates.append(op.predicates)
-        elif isinstance(right, ops.Join):
-            # When rewrites are possible at the expression IR stage, we should
-            # do them. Otherwise subqueries might be necessary in some cases
-            # here
-            raise NotImplementedError('not allowing joins on right '
-                                      'side yet')
-        else:
-            # Both tables
-            self.join_tables.append(self._format_table(op.left))
-            self.join_tables.append(self._format_table(op.right))
-            self.join_types.append(jname)
-            self.join_predicates.append(op.predicates)
-
-    def _get_join_type(self, op):
-        jname = self._join_names[type(op)]
-
-        # Impala requires this
-        if len(op.predicates) == 0:
-            jname = self._join_names[ops.CrossJoin]
-
-        return jname
-
-    def _format_table(self, expr):
-        # TODO: This could probably go in a class and be significantly nicer
-        from ibis.impala.exprs import quote_identifier
-
-        ctx = self.context
-
-        ref_expr = expr
-        op = ref_op = expr.op()
-        if isinstance(op, ops.SelfReference):
-            ref_expr = op.table
-            ref_op = ref_expr.op()
-
-        if isinstance(ref_op, ops.PhysicalTable):
-            name = ref_op.name
-            if name is None:
-                raise com.RelationError('Table did not have a name: {0!r}'
-                                        .format(expr))
-            result = quote_identifier(name)
-            is_subquery = False
-        else:
-            # A subquery
-            if ctx.is_extracted(ref_expr):
-                # Was put elsewhere, e.g. WITH block, we just need to grab its
-                # alias
-                alias = ctx.get_ref(expr)
-
-                # HACK: self-references have to be treated more carefully here
-                if isinstance(op, ops.SelfReference):
-                    return '{0} {1}'.format(ctx.get_ref(ref_expr), alias)
-                else:
-                    return alias
-
-            subquery = ctx.get_compiled_expr(expr)
-            result = '(\n{0}\n)'.format(util.indent(subquery, self.indent))
-            is_subquery = True
-
-        if is_subquery or ctx.need_aliases():
-            result += ' {0}'.format(ctx.get_ref(expr))
-
-        return result
-
-    # Placeholder; revisit when supporting other databases
-    _non_equijoin_supported = True
-
-    def _validate_join_predicates(self, predicates):
-        for pred in predicates:
-            op = pred.op()
-
-            if (not isinstance(op, ops.Equals) and
-                    not self._non_equijoin_supported):
-                raise com.TranslationError(
-                    'Non-equality join predicates, '
-                    'i.e. non-equijoins, are not supported')
-
-
-class Union(DDL):
-
-    def __init__(self, left_table, right_table, distinct=False,
-                 context=None):
-        self.context = context
-        self.left = left_table
-        self.right = right_table
-
-        self.distinct = distinct
-
-    def compile(self):
-        context = self.context
-
-        if self.distinct:
-            union_keyword = 'UNION'
-        else:
-            union_keyword = 'UNION ALL'
-
-        left_set = context.get_compiled_expr(self.left)
-        right_set = context.get_compiled_expr(self.right)
-
-        query = '{0}\n{1}\n{2}'.format(left_set, union_keyword, right_set)
-        return query
-
-
-class ExprTranslator(object):
-
-    def __init__(self, expr, context=None, named=False, permit_subquery=False):
-        self.expr = expr
-        self.permit_subquery = permit_subquery
-
-        if context is None:
-            context = self._context_class()
-        self.context = context
-
-        # For now, governing whether the result will have a name
-        self.named = named
-
-    @property
-    def _context_class(self):
-        from ibis.sql.compiler import QueryContext
-        return QueryContext
-
-    def get_result(self):
-        """
-        Build compiled SQL expression from the bottom up and return as a string
-        """
-        translated = self.translate(self.expr)
-        if self._needs_name(self.expr):
-            # TODO: this could fail in various ways
-            name = self.expr.get_name()
-            translated = self.name(translated, name)
-        return translated
-
-    def _needs_name(self, expr):
-        if not self.named:
-            return False
-
-        op = expr.op()
-        if isinstance(op, ops.TableColumn):
-            # This column has been given an explicitly different name
-            if expr.get_name() != op.name:
-                return True
-            return False
-
-        if expr.get_name() is ir.unnamed:
-            return False
-
-        return True
-
-    def translate(self, expr):
-        # The operation node type the typed expression wraps
-        op = expr.op()
-
-        if type(op) in self._rewrites:
-            expr = self._rewrites[type(op)](expr)
-            op = expr.op()
-
-        # TODO: use op MRO for subclasses instead of this isinstance spaghetti
-        if isinstance(op, ir.Parameter):
-            return self._trans_param(expr)
-        elif isinstance(op, ops.TableNode):
-            # HACK/TODO: revisit for more complex cases
-            return '*'
-        elif type(op) in self._registry:
-            formatter = self._registry[type(op)]
-            return formatter(self, expr)
-        else:
-            raise com.TranslationError('No translator rule for {0}'.format(op))
-
-    def _trans_param(self, expr):
-        raise NotImplementedError
-
-
-def _join_not_none(sep, pieces):
-    pieces = [x for x in pieces if x is not None]
-    return sep.join(pieces)
