+++ ibis/sql/sqlite/client.py
@@ -12,11 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ibis.client import SQLClient
 import ibis.expr.types as ir
 import ibis.sql.alchemy as alch
-import sqlalchemy as sa
-
 from .compiler import SQLiteDialect
